requirement.txt
cmake
numpy
opencv-python
keras
Pillow
tensorflow
pyttsx3==2.9.0